黑马头条PC-发布文章页面排版布局

+++ src/views/contentList/contentList.vue
@@ -244,22 +244,37 @@ export default {
     },
     // 点击删除按钮触发的方法
     onDelete (articleID) {
-      console.log('ID是' + articleID)
+      // console.log('ID是' + articleID)
       const token = window.localStorage.getItem('token')
-      // ID获取到了，发送请求就OK啦
-      this.$axios({
+      this.$confirm('确定要删除该文章吗, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        // ID获取到了，发送请求就OK啦
+        this.$axios({
         // 请求地址
-        url: `/articles/${articleID}`,
-        // 请求方式
-        method: 'DELETE',
-        // 在请求请求头中携带token
-        headers: {
-          Authorization: `Bearer ${token}`
-        }
-      }).then(res => {
-        console.log(res)
-      }).catch(erro => {
-        console.log(erro, '删除貌似失败了')
+          url: `/articles/${articleID}`,
+          // 请求方式
+          method: 'DELETE',
+          // 在请求请求头中携带token
+          headers: {
+            Authorization: `Bearer ${token}`
+          }
+        }).then(res => {
+          console.log(res)
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          })
+        }).catch(erro => {
+          console.log(erro, '删除貌似失败了')
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
     }
   },
